+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Settings - Usage Consolidation - Currency. (UIEH-965)
 * Settings - Usage Consolidation - Start month for usage statistics. (UIEH-966)
 * Settings - Usage Consolidation - Usage consolidation ID. (UIEH-968)
+* Added Usage Consolidation Filters to Package Show page. (UIEH-942, UIEH-973)
 
 ## [5.0.0] (https://github.com/folio-org/ui-eholdings/tree/v5.0.0) (2020-10-15)